analyze-expense-refactor
